requirements.txt
Flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.0.33
scikit-learn==1.3.0
nltk==3.8.1